requirements.txt
django==5.0.1
djangorestframework==3.14.0
pytest-django==4.8.0
factory-boy==3.3.0